remove shared csproj and fix -f
