+++ src/Content/BlazorMaui/BlazorMaui.Rcl/_Imports.razor
@@ -3,4 +3,3 @@
 @using BlazorComponent
 @using Masa.Blazor
 @using BlazorMaui.Rcl.Shared
-@using BlazorMaui.Shared
