+++ src/Content/BlazorMaui/BlazorMaui.Server/Program.cs
@@ -1,4 +1,4 @@
-using BlazorMaui.Shared;
+using BlazorMaui.Rcl;
 
 var builder = WebApplication.CreateBuilder(args);
