--- src/Content/BlazorMaui/BlazorMaui.Shared/WeatherForecast.cs
@@ -1,12 +1,0 @@
-namespace BlazorMaui.Shared;
-
-public class WeatherForecast
-{
-    public DateOnly Date { get; set; }
-
-    public int TemperatureC { get; set; }
-
-    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-
-    public string? Summary { get; set; }
-}
