+++ src/Content/BlazorMaui/BlazorMaui/MauiProgram.cs
@@ -1,4 +1,4 @@
-﻿using BlazorMaui.Shared;
+﻿using BlazorMaui.Rcl;
 using Microsoft.Extensions.Logging;
 
 namespace BlazorMaui;
