+++ src/Content/BlazorMaui/BlazorMaui/_Imports.razor
@@ -5,5 +5,4 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
 @using BlazorMaui
-@using BlazorMaui.Shared
 @using BlazorMaui.Rcl.Shared
