+++ src/Content/BlazorMaui/BlazorMaui/wwwroot/css/app.css
@@ -1,4 +1,12 @@
-﻿#blazor-error-ui {
+﻿html, body {
+    font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
+}
+
+h1:focus {
+    outline: none;
+}
+
+#blazor-error-ui {
     background: lightyellow;
     bottom: 0;
     box-shadow: 0 -1px 2px rgba(0, 0, 0, 0.2);
